Finally added some simple (crap) physics.
It isn't my best work and the design's slightly unintuitive, but it works and it should work with furthur additions.
That being said, this commit is nothing to celebrate, The ball only moves up and down and doesn't even work right since the collision code doesn't factor in the paddle's x position.