GameSettings with JSON have been implemented!
Specifically, the function pong::loadSettings has been added which does
all the parsing on a passed in filename. This implementation has not
been used yet though...

In order to parse JSON, jsoncpp has been added to the source tree.

The original source code can be found at
sourceforge: http://sourceforge.net/projects/jsoncpp/

The source code distributed here was generated with the provided
amalgamate.py script and the LICENSE was copied from the source tree
(of jsoncpp, that is).