Modernized the SinglePlayerGameState.
The "SinglePlayer" button has also been enabled in the menu. Currently
though, there is a bug where after pressing a key you will return to the
menu but the mouse will not be freed or shown, that is because the
MenuGameState is not constructed again. It is time to add a couple of
functions along the lines of onFocus and onUnfocus.