Cached the GameState in the main loop to prevent weird behavior.
If the game state is changed by another gamestate. The game loop needs
to know that, so it can restart the loop. The way it does this
currently, is by making another shared pointer to share ownership with
the Game class's one. This way, not only is the game state guaranteed to
stay alive, but a change can be detected by comparing it against the
Game class's current game_state.