Small change in the order of actions taken by Label's move operator=.
By moving the cache move to the end we counteract any of the members
initiative to destroy the cache, since it doesn't no better. In other
words, the changes to the members will most likely cause an invalidation
of the cache, because the good, alive and awesome one was moved in
already *it* is going to be the one to be invalidated. By moving it to
the end the most a member setter is going to do is free the surface we
don't want anymore.