Simplified interface in ControlInterface.h as well as complete reimplementation, sorta!
Basically three separately initialized pipes are used for a process'
stdin, stdout and stderr. This allows for easy i/o with a guarantee that
a message is being sent to the write process and not being
misinterpretted by another, etc.

It is also has some policy implemented which needs to be taken out soon.
For now though LocalServer creates a child 'cat' process which is sent a
single message (the text 'PPM') then given EOF.