Changed the default font colors to more anti-aliased friendly values.
By interpolating between the same color with opposite alpha values we get
a much better transition where it matters. For instance when text is
being rendered around or on top of other objects of the same color.
Previously these labels' colors were interpolated (not really because it
was monochrome) between opaque white and transparent black which looks
weird when rendering grayscale text.