Response methods are not added in an external function.
Lot of separation of code between engine.cpp and methods.cpp with
engine.h holding the common stuff and structs, etc.

The Req_Method and Req_Method_Base now return a response_result which is
just the variant type of the Response's result member. This way methods
don't have to make up an id that is going to be chucked anyway.