The engine now just lets the dispatcher Do-The-Right-Thing.
That is, it no longer adds a null_t value type to the id of an error
response, since ideally, only some errors are returned and that's only
if they were serious enough for the Req_Dispatcher to make them null_t
in the first place. Otherwise we don't want to spit out every error to
every notification, for instance.